完成作业人员资格证书生成

+++ continew-webapi/src/main/java/top/continew/admin/exam/model/dto/CheckPlanHasExamTypeDTO.java
@@ -12,7 +12,7 @@ public class CheckPlanHasExamTypeDTO {
     /**
      * 计划id
      */
-    private Integer planId;
+    private Long planId;
 
     /**
      * 计划名称
